detection

+++ torchTools/detectionv2/datasets/bboxAug.py
@@ -105,14 +105,14 @@ class Pad(object):
             pad_list = [[0, 0], [diff // 2, diff - diff // 2], [0, 0]]
             if "boxes" in target:
                 boxes = [[b[0] + diff // 2, b[1], b[2] + diff - diff // 2, b[3]] for b in boxes]
-                boxes = torch.as_tensor(boxes)
+                boxes = torch.as_tensor(boxes,dtype=torch.float32)
 
         else:
             diff = w - h
             pad_list = [[diff // 2, diff - diff // 2], [0, 0], [0, 0]]
             if "boxes" in target:
                 boxes = [[b[0], b[1] + diff // 2, b[2], b[3] + diff - diff // 2] for b in boxes]
-                boxes = torch.as_tensor(boxes)
+                boxes = torch.as_tensor(boxes,dtype=torch.float32)
 
         img = np.pad(img, pad_list, mode=self.mode, constant_values=self.value)
 
@@ -318,7 +318,7 @@ class RandomHorizontalFlip(object):
         img = np.asarray(img)
         img_center = np.array(img.shape[:2])[::-1] / 2
         img_center = np.hstack((img_center, img_center))
-        img_center = torch.as_tensor(img_center)
+        img_center = torch.as_tensor(img_center,dtype=torch.float32)
         bboxes = target["boxes"]
         if random.random() < self.p:
             img = img[:, ::-1, :]
@@ -344,7 +344,7 @@ class RandomVerticallyFlip(object):
         img = np.asarray(img)
         img_center = np.array(img.shape[:2])[::-1] / 2
         img_center = np.hstack((img_center, img_center))
-        img_center = torch.as_tensor(img_center)
+        img_center = torch.as_tensor(img_center,dtype=torch.float32)
         bboxes = target["boxes"]
         if random.random() < self.p:
             img = img[::-1, :, :]
@@ -404,7 +404,7 @@ class RandomScale(object):
         img = canvas
         bboxes = clip_box(bboxes, [0, 0, 1 + img_shape[1], img_shape[0]], 0.25)
 
-        target["boxes"] = torch.as_tensor(bboxes)
+        target["boxes"] = torch.as_tensor(bboxes,dtype=torch.float32)
 
         return PIL.Image.fromarray(img), target
 
@@ -425,7 +425,7 @@ class RandomScale2(object):
             scale_tensor = torch.FloatTensor([[self.scale, 1, self.scale, 1]]).expand_as(boxes)
             boxes = boxes * scale_tensor
 
-            target["boxes"] = torch.as_tensor(boxes)
+            target["boxes"] = torch.as_tensor(boxes,dtype=torch.float32)
             img = PIL.Image.fromarray(img)
 
         return img, target
@@ -481,7 +481,7 @@ class RandomTranslate(object):
 
         bboxes = clip_box(bboxes, [0, 0, img_shape[1], img_shape[0]], 0.25)
 
-        target["boxes"] = torch.as_tensor(bboxes)
+        target["boxes"] = torch.as_tensor(bboxes,dtype=torch.float32)
 
         return PIL.Image.fromarray(img), target
 
@@ -529,7 +529,7 @@ class RandomRotate(object):
 
         bboxes = clip_box(bboxes, [0, 0, w, h], 0.25)
 
-        target["boxes"] = torch.as_tensor(bboxes)
+        target["boxes"] = torch.as_tensor(bboxes,dtype=torch.float32)
 
         return PIL.Image.fromarray(img), target
